add method working
loop that continues until done becomes
true. Each iteration of the loop handles these two cases:
1. Suppose the new element is less than or equal to the data at the
cursor
node. In this case, check the left child. If cursor.getLeft( ) is null,
then
create a new node containing the new element (setting the cursor’s left
child to this new node) and set done to true. On the other hand, if the
left child is not null, then continue the search by moving the cursor to
the left with this assignment statement:
cursor = cursor.getLeft( );
2. Suppose the new element is greater than the data at the cursor node.
In
this case, follow the same procedure as Case 1, but use the right child
instead of the left.
After the loop ends, the method has no more work.